feat(communicationmapping): move to ied-editor
